requirements.txt
tqdm==4.48.0
warcio==1.7.3
requests==2.24.0
lm_dataformat==0.0.8
cchardet==2.1.6
jsonlines==1.2.0
pycld2==0.41
justext==2.2.0
fasttext==0.9.2
trafilatura==0.5.1
pybloomfiltermmap3
loguru
ray==1.2.0
